Create interface for VnodeRepairStates
