+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/repair/state/VnodeRepairStateFactoryImpl.java
@@ -79,7 +79,7 @@ public class VnodeRepairStateFactoryImpl implements VnodeRepairStateFactory
             vnodeRepairStatesBase.add(new VnodeRepairState(longTokenRange, replicas, lastRepairedAt));
         }
 
-        VnodeRepairStates.Builder vnodeRepairStatusesBuilder = VnodeRepairStates.newBuilder(vnodeRepairStatesBase);
+        VnodeRepairStates.Builder vnodeRepairStatusesBuilder = VnodeRepairStatesImpl.newBuilder(vnodeRepairStatesBase);
 
         if (previous != null)
         {
