+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/repair/state/VnodeRepairStates.java
@@ -14,30 +14,15 @@
  */
 package com.ericsson.bss.cassandra.ecchronos.core.repair.state;
 
-import com.ericsson.bss.cassandra.ecchronos.core.utils.LongTokenRange;
-import com.google.common.collect.ImmutableList;
-
 import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Objects;
 
 /**
  * A collection of {@link VnodeRepairState VnodeRepairStates} that contains utilities to easily combine with new entries.
  */
-public class VnodeRepairStates
+public interface VnodeRepairStates
 {
-    private final ImmutableList<VnodeRepairState> myVnodeRepairStatuses;
 
-    private VnodeRepairStates(Builder builder)
-    {
-        myVnodeRepairStatuses = ImmutableList.copyOf(builder.myVnodeRepairStates.values());
-    }
-
-    public Collection<VnodeRepairState> getVnodeRepairStates()
-    {
-        return myVnodeRepairStatuses;
-    }
+    Collection<VnodeRepairState> getVnodeRepairStates();
 
     /**
      * Create a new vnode repair states object with the minimum repaired at set to the provided value.
@@ -47,57 +32,10 @@ public class VnodeRepairStates
      * @param repairedAt The minimum repaired at to use.
      * @return The created state.
      */
-    public VnodeRepairStates combineWithRepairedAt(long repairedAt)
+    VnodeRepairStates combineWithRepairedAt(long repairedAt);
+
+    interface Builder
     {
-        Builder builder = newBuilder(getVnodeRepairStates());
-
-        for (VnodeRepairState vnodeRepairState : getVnodeRepairStates())
-        {
-            VnodeRepairState vnodeRepairStateWithRepairedAt = new VnodeRepairState(vnodeRepairState.getTokenRange(), vnodeRepairState.getReplicas(), repairedAt);
-            builder.updateVnodeRepairState(vnodeRepairStateWithRepairedAt);
-        }
-
-        return builder.build();
-    }
-
-    @Override
-    public String toString()
-    {
-        return myVnodeRepairStatuses.toString();
-    }
-
-    public static Builder newBuilder(Collection<VnodeRepairState> vnodeRepairStates)
-    {
-        return new Builder(vnodeRepairStates);
-    }
-
-    @Override
-    public boolean equals(Object o)
-    {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        VnodeRepairStates that = (VnodeRepairStates) o;
-        return Objects.equals(myVnodeRepairStatuses, that.myVnodeRepairStatuses);
-    }
-
-    @Override
-    public int hashCode()
-    {
-        return Objects.hash(myVnodeRepairStatuses);
-    }
-
-    public static class Builder
-    {
-        private final Map<LongTokenRange, VnodeRepairState> myVnodeRepairStates = new HashMap<>();
-
-        public Builder(Collection<VnodeRepairState> vnodeRepairStates)
-        {
-            for (VnodeRepairState vnodeRepairState : vnodeRepairStates)
-            {
-                myVnodeRepairStates.put(vnodeRepairState.getTokenRange(), vnodeRepairState);
-            }
-        }
-
         /**
          * Combine a collection of vnode repair states into this collection.
          *
@@ -105,7 +43,7 @@ public class VnodeRepairStates
          * @return This builder
          * @see #updateVnodeRepairState(VnodeRepairState)
          */
-        public Builder updateVnodeRepairStates(Collection<VnodeRepairState> vnodeRepairStates)
+        default Builder updateVnodeRepairStates(Collection<VnodeRepairState> vnodeRepairStates)
         {
             for (VnodeRepairState vnodeRepairState : vnodeRepairStates)
             {
@@ -124,34 +62,8 @@ public class VnodeRepairStates
          * @param vnodeRepairState The vnode repair status to update.
          * @return This builder
          */
-        public Builder updateVnodeRepairState(VnodeRepairState vnodeRepairState)
-        {
-            VnodeRepairState oldVnode = myVnodeRepairStates.get(vnodeRepairState.getTokenRange());
-            if (shouldReplace(oldVnode, vnodeRepairState))
-            {
-                myVnodeRepairStates.put(vnodeRepairState.getTokenRange(), vnodeRepairState);
-            }
-            return this;
-        }
+        Builder updateVnodeRepairState(VnodeRepairState vnodeRepairState);
 
-        private boolean shouldReplace(VnodeRepairState oldVnode, VnodeRepairState newVnode)
-        {
-            if (oldVnode == null)
-            {
-                return false;
-            }
-
-            if (!oldVnode.isSameVnode(newVnode))
-            {
-                return false;
-            }
-
-            return oldVnode.lastRepairedAt() < newVnode.lastRepairedAt();
-        }
-
-        public VnodeRepairStates build()
-        {
-            return new VnodeRepairStates(this);
-        }
+        VnodeRepairStates build();
     }
 }
